restructure jinja plugin as a sub package
