--- development/coverage_plugins/coverage_plugins/jinja.py
@@ -1,1000 +1,0 @@
-# Copyright (c) 2023-2026 Arista Networks, Inc.
-# Use of this source code is governed by the Apache License 2.0
-# that can be found in the LICENSE file.
-"""
-Coverage.py plugin mapping generated Jinja module execution to source templates.
-
-The reporting model is intentionally source-level and heuristic: it maps only
-lines and arcs it can associate with template behavior instead of assigning
-generated scaffolding to the nearest template line.
-"""
-
-from __future__ import annotations
-
-import ast
-from dataclasses import dataclass
-from functools import lru_cache
-from pathlib import Path
-from types import MappingProxyType
-from typing import TYPE_CHECKING, Any
-
-from coverage import CoveragePlugin, FileReporter, FileTracer
-from coverage.exceptions import ConfigError
-
-if TYPE_CHECKING:
-    from collections.abc import Collection, Iterable, Mapping
-    from types import FrameType
-
-JINJA2_EXTENSIONS = ("jinja2.ext.loopcontrols", "jinja2.ext.do", "jinja2.ext.i18n")
-SOURCE_EXIT = -1
-FileStamp = tuple[int, int]
-
-
-@dataclass(frozen=True)
-class CompiledTemplate:
-    """Parsed mapping metadata for one compiled Jinja Python module."""
-
-    source_filename: str
-    debug_map: tuple[tuple[int, int], ...]
-    generated_line_ranges: Mapping[int, tuple[int, int]]
-
-
-@dataclass(frozen=True)
-class SourceTemplate:
-    """Cached source-level reporting model for one Jinja template."""
-
-    reportable_lines: frozenset[int]
-    arc_endpoint_lines: frozenset[int]
-    possible_arcs: frozenset[tuple[int, int]]
-    no_branch_lines: frozenset[int]
-    tag_ranges: Mapping[int, tuple[int, int]]
-
-
-class JinjaTemplateCoveragePlugin(CoveragePlugin):
-    """Coverage.py file tracer plugin for generated Jinja modules."""
-
-    def __init__(self, compiled_template_roots: Iterable[str | Path] | None = None) -> None:
-        """Store normalized directories containing generated Jinja Python modules."""
-        self.compiled_template_roots = _configured_compiled_template_roots(compiled_template_roots)
-
-    def file_tracer(self, filename: str) -> JinjaTemplateFileTracer | None:
-        """Return a tracer for generated Jinja modules under configured compiled-template roots."""
-        compiled_filename = Path(filename).resolve()
-        if compiled_filename.suffix != ".py":
-            return None
-
-        compiled_root = self._compiled_root_for(compiled_filename)
-        if compiled_root is None:
-            return None
-
-        compiled_template = _parse_compiled_template(compiled_filename, compiled_root)
-        if compiled_template is None:
-            return None
-
-        return JinjaTemplateFileTracer(compiled_template)
-
-    def file_reporter(self, filename: str) -> JinjaTemplateFileReporter:
-        """Return the reporter used by coverage.py when generating reports for a source template."""
-        return JinjaTemplateFileReporter(filename)
-
-    def _compiled_root_for(self, filename: Path) -> Path | None:
-        """Find the configured compiled-template root containing ``filename``."""
-        for compiled_root in self.compiled_template_roots:
-            if filename.is_relative_to(compiled_root):
-                return compiled_root
-        return None
-
-
-class JinjaTemplateFileTracer(FileTracer):
-    """Map executed generated Python frame lines to source Jinja template lines."""
-
-    def __init__(self, compiled_template: CompiledTemplate) -> None:
-        """Keep the precomputed source mapping for one compiled template module."""
-        self.compiled_template = compiled_template
-
-    def source_filename(self) -> str:
-        """Tell coverage.py which source file should receive execution data."""
-        return self.compiled_template.source_filename
-
-    def line_number_range(self, frame: FrameType) -> tuple[int, int]:
-        """Translate the current generated Python frame line to a source template line range."""
-        generated_line = frame.f_lineno
-        if line_range := self.compiled_template.generated_line_ranges.get(generated_line):
-            return line_range
-
-        return -1, -1
-
-
-class JinjaTemplateFileReporter(FileReporter):
-    """Describe executable lines and branch arcs for a source Jinja template."""
-
-    def lines(self) -> set[int]:
-        """Return source template lines that should be treated as executable statements."""
-        return set(_source_template(Path(self.filename)).reportable_lines)
-
-    def translate_lines(self, lines: Iterable[int]) -> set[int]:
-        """Drop recorded lines that are not reportable source template lines."""
-        reportable_lines = self.lines()
-        return {line for line in lines if line in reportable_lines}
-
-    def arcs(self) -> set[tuple[int, int]]:
-        """Return possible source-level branch arcs for supported Jinja control flow."""
-        return set(_source_template(Path(self.filename)).possible_arcs)
-
-    def no_branch_lines(self) -> set[int]:
-        """Return lines that should not be treated as missing branch coverage."""
-        return set(_source_template(Path(self.filename)).no_branch_lines)
-
-    def translate_arcs(self, arcs: Iterable[tuple[int, int]]) -> set[tuple[int, int]]:
-        """Drop recorded arcs whose endpoints are not reportable source template lines."""
-        recorded_arcs = tuple(arcs)
-        source_template = _source_template(Path(self.filename))
-        translated_arcs: set[tuple[int, int]] = set()
-        for from_line, to_line in recorded_arcs:
-            translated_from_line = _translate_recorded_arc_endpoint(from_line, source_template.arc_endpoint_lines)
-            translated_to_line = _translate_recorded_arc_endpoint(to_line, source_template.arc_endpoint_lines)
-            if translated_from_line is not None and translated_to_line is not None:
-                translated_arcs.add((translated_from_line, translated_to_line))
-
-        possible_arcs = source_template.possible_arcs
-        source_filename = Path(self.filename)
-        translated_arcs.update(
-            _covered_multiline_tag_branch_arcs(recorded_arcs, possible_arcs, source_template.tag_ranges, source_template.reportable_lines),
-        )
-        translated_arcs.update(
-            _covered_else_branch_arcs(recorded_arcs, translated_arcs, possible_arcs, source_filename, source_template.reportable_lines),
-        )
-        return translated_arcs
-
-    def exit_counts(self) -> dict[int, int]:
-        """Return the number of possible exits from each reportable source template line."""
-        exit_counts: dict[int, int] = {}
-        for from_line, to_line in self.arcs():
-            if from_line > 0 and to_line != from_line:
-                exit_counts[from_line] = exit_counts.get(from_line, 0) + 1
-
-        return exit_counts
-
-
-def coverage_init(reg, options) -> None:  # noqa: ANN001
-    """Register the plugin with coverage.py using options from ``[tool.coverage.coverage_plugins.jinja]``."""
-    configured_roots = options.get("compiled_template_roots")
-    if not configured_roots:
-        msg = "coverage_plugins.jinja requires the compiled_template_roots coverage plugin option."
-        raise ConfigError(msg)
-
-    reg.add_file_tracer(JinjaTemplateCoveragePlugin(compiled_template_roots=configured_roots))
-
-
-def _configured_compiled_template_roots(configured_roots: Iterable[str | Path] | str | Path | None) -> tuple[Path, ...]:
-    """Normalize the configured compiled-template roots to absolute ``Path`` objects."""
-    if configured_roots is None:
-        return ()
-
-    if isinstance(configured_roots, (str, Path)):
-        configured_roots = (configured_roots,)
-
-    return tuple(Path(root).resolve() for root in configured_roots)
-
-
-def _parse_compiled_template(compiled_filename: Path, compiled_root: Path) -> CompiledTemplate | None:
-    """
-    Resolve and parse a compiled Jinja Python module with stamp-based cache invalidation.
-
-    The mapping result depends on both the generated Python module and the
-    source ``.j2`` template. The wrapper reads file stamps first and passes
-    those stamps into cached helpers so edits to either file invalidate the
-    cached mapping.
-    """
-    compiled_filename = compiled_filename.resolve()
-    compiled_root = compiled_root.resolve()
-    if (compiled_stamp := _file_stamp(compiled_filename)) is None:
-        return None
-
-    source_filename = _resolve_compiled_template_source_filename(compiled_filename, compiled_root, compiled_stamp)
-    if source_filename is None:
-        return None
-
-    if (source_stamp := _file_stamp(source_filename)) is None:
-        return None
-
-    return _parse_compiled_template_cached(compiled_filename, source_filename.resolve(), compiled_stamp, source_stamp)
-
-
-@lru_cache(maxsize=4096)
-def _parse_compiled_template_cached(
-    compiled_filename: Path,
-    source_filename: Path,
-    compiled_stamp: FileStamp,  # noqa: ARG001
-    source_stamp: FileStamp,  # noqa: ARG001
-) -> CompiledTemplate | None:
-    """
-    Build immutable line mappings for one compiled template.
-
-    ``compiled_stamp`` and ``source_stamp`` are intentionally unused inside the
-    function body. They are part of the cache key so repeated tracing of the
-    same generated module is cheap, while file rewrites still invalidate stale
-    mappings.
-    """
-    try:
-        tree = ast.parse(compiled_filename.read_text(encoding="utf-8"), filename=str(compiled_filename))
-    except (OSError, SyntaxError, UnicodeDecodeError):
-        return None
-
-    constants = _module_string_constants(tree)
-    debug_info = constants.get("debug_info")
-    if debug_info is None:
-        return None
-
-    debug_map = _parse_debug_info(debug_info)
-    generated_line_ranges = _generated_line_ranges(tree, source_filename, debug_map)
-    return CompiledTemplate(
-        source_filename=str(source_filename),
-        debug_map=debug_map,
-        generated_line_ranges=MappingProxyType(generated_line_ranges),
-    )
-
-
-def _generated_line_ranges(tree: ast.Module, source_filename: Path, debug_map: tuple[tuple[int, int], ...]) -> dict[int, tuple[int, int]]:
-    """Return generated-line mappings for Jinja tag ranges and static output ranges."""
-    source_tag_ranges = _source_tag_ranges(source_filename)
-    else_branch_line_ranges = _else_branch_line_ranges(source_filename)
-    generated_line_ranges = {
-        generated_line: source_tag_ranges.get(template_line, (template_line, template_line)) for generated_line, template_line in debug_map
-    }
-    generated_line_ranges.update(_generated_static_line_ranges(tree, source_filename))
-    generated_line_ranges.update(_generated_endif_line_ranges(tree, source_filename, debug_map, generated_line_ranges))
-    for generated_line, (start_line, end_line) in generated_line_ranges.items():
-        if else_branch_range := else_branch_line_ranges.get(start_line):
-            generated_line_ranges[generated_line] = (else_branch_range[0], end_line)
-
-    return generated_line_ranges
-
-
-@lru_cache(maxsize=4096)
-def _resolve_compiled_template_source_filename(compiled_filename: Path, compiled_root: Path, compiled_stamp: FileStamp) -> Path | None:  # noqa: ARG001
-    """
-    Read a compiled module's ``name`` constant and resolve it to a source template path.
-
-    The generated Jinja module records the original template name in a module
-    constant. ``compiled_stamp`` is included only as cache invalidation input.
-    """
-    try:
-        tree = ast.parse(compiled_filename.read_text(encoding="utf-8"), filename=str(compiled_filename))
-    except (OSError, SyntaxError, UnicodeDecodeError):
-        return None
-
-    constants = _module_string_constants(tree)
-    template_name = constants.get("name")
-    if template_name is None or constants.get("debug_info") is None:
-        return None
-
-    return _resolve_template_source_filename(template_name, compiled_root)
-
-
-def _source_template(filename: Path) -> SourceTemplate:
-    """Return the cached source reporting model for one template."""
-    if (file_stamp := _file_stamp(filename)) is None:
-        return SourceTemplate(
-            reportable_lines=frozenset(),
-            arc_endpoint_lines=frozenset(),
-            possible_arcs=frozenset(),
-            no_branch_lines=frozenset(),
-            tag_ranges=MappingProxyType({}),
-        )
-
-    return _source_template_cached(filename.resolve(), file_stamp)
-
-
-@lru_cache(maxsize=4096)
-def _source_template_cached(filename: Path, file_stamp: FileStamp) -> SourceTemplate:
-    """
-    Build the source-level model used by all reporting methods.
-
-    The arc endpoint set is explicit so possible arc generation and recorded
-    arc translation agree on which source lines can be branch endpoints.
-    ``file_stamp`` participates only in cache invalidation.
-    """
-    reportable_lines = _find_reportable_jinja_lines_cached(filename, file_stamp)
-    structural_control_label_lines = _find_structural_control_label_lines_cached(filename, file_stamp)
-    arc_endpoint_lines = _arc_endpoint_lines(reportable_lines, structural_control_label_lines)
-    return SourceTemplate(
-        reportable_lines=reportable_lines,
-        arc_endpoint_lines=arc_endpoint_lines,
-        possible_arcs=_find_possible_jinja_arcs_cached(filename, file_stamp, reportable_lines, arc_endpoint_lines),
-        no_branch_lines=_find_no_branch_jinja_lines_cached(filename, file_stamp),
-        tag_ranges=MappingProxyType(_source_tag_ranges(filename)),
-    )
-
-
-def _arc_endpoint_lines(reportable_lines: Collection[int], structural_control_label_lines: Collection[int] = ()) -> frozenset[int]:
-    """
-    Return source lines allowed as branch arc endpoints.
-
-    Structural control labels are not executable lines, but they are useful
-    source-level join points for branch coverage.
-    """
-    return frozenset(reportable_lines) | frozenset(structural_control_label_lines)
-
-
-def _translate_recorded_arc_endpoint(line: int, arc_endpoint_lines: Collection[int]) -> int | None:
-    """Translate one recorded arc endpoint to a template branch endpoint."""
-    if line < 0:
-        return SOURCE_EXIT
-
-    if line in arc_endpoint_lines:
-        return line
-
-    return None
-
-
-def _covered_multiline_tag_branch_arcs(
-    recorded_arcs: tuple[tuple[int, int], ...],
-    possible_arcs: Collection[tuple[int, int]],
-    tag_ranges: Mapping[int, tuple[int, int]],
-    reportable_lines: Collection[int],
-) -> set[tuple[int, int]]:
-    """Return source branch arcs covered by coverage.py's multiline range arcs."""
-    recorded_arc_set = set(recorded_arcs)
-    covered_branch_arcs: set[tuple[int, int]] = set()
-    for from_line, to_line in possible_arcs:
-        if from_line <= 0 or to_line <= 0:
-            continue
-        if to_line not in reportable_lines:
-            continue
-
-        tag_range = tag_ranges.get(from_line)
-        if tag_range is None or tag_range[0] == tag_range[1]:
-            continue
-
-        start_line, end_line = tag_range
-        if not all((line_number, line_number + 1) in recorded_arc_set for line_number in range(start_line, end_line)):
-            continue
-
-        if any(raw_to_line == to_line and raw_from_line not in reportable_lines for raw_from_line, raw_to_line in recorded_arcs):
-            covered_branch_arcs.add((from_line, to_line))
-
-    return covered_branch_arcs
-
-
-def _covered_else_branch_arcs(
-    recorded_arcs: tuple[tuple[int, int], ...],
-    translated_arcs: set[tuple[int, int]],
-    possible_arcs: Collection[tuple[int, int]],
-    source_filename: Path,
-    reportable_lines: Collection[int],
-) -> set[tuple[int, int]]:
-    """Return else branch arcs covered by generated arcs entering an else tag."""
-    try:
-        source = source_filename.read_text(encoding="utf-8")
-    except (OSError, UnicodeDecodeError):
-        return set()
-
-    else_lines = _control_statement_lines(source, names={"else"})
-    executed_lines = {line for arc in translated_arcs for line in arc if line > 0}
-    covered_else_arcs: set[tuple[int, int]] = set()
-    for from_line, to_line in possible_arcs:
-        if from_line not in executed_lines or to_line not in else_lines:
-            continue
-
-        if any(raw_to_line == to_line and raw_from_line not in reportable_lines for raw_from_line, raw_to_line in recorded_arcs):
-            covered_else_arcs.add((from_line, to_line))
-
-    return covered_else_arcs
-
-
-def _generated_endif_line_ranges(
-    tree: ast.Module,
-    source_filename: Path,
-    debug_map: tuple[tuple[int, int], ...],
-    existing_line_ranges: Mapping[int, tuple[int, int]],
-) -> dict[int, tuple[int, int]]:
-    """Map generated post-if statements to the matching source ``endif`` label."""
-    try:
-        source = source_filename.read_text(encoding="utf-8")
-    except (OSError, UnicodeDecodeError):
-        return {}
-
-    endif_lines_by_conditional_line = _if_endif_lines(source)
-    if not endif_lines_by_conditional_line:
-        return {}
-
-    source_lines_by_generated_line = dict(debug_map)
-    endif_ranges: dict[int, tuple[int, int]] = {}
-
-    def visit_statements(statements: list[ast.stmt]) -> None:
-        for index, statement in enumerate(statements):
-            source_line = source_lines_by_generated_line.get(statement.lineno) if isinstance(statement, ast.If) else None
-            if source_line is not None and (endif_line := endif_lines_by_conditional_line.get(source_line)) is not None and index + 1 < len(statements):
-                next_generated_line = statements[index + 1].lineno
-                existing_start_line, existing_end_line = existing_line_ranges.get(next_generated_line, (endif_line, endif_line))
-                endif_ranges[next_generated_line] = (min(endif_line, existing_start_line), max(endif_line, existing_end_line))
-
-            for _field_name, value in ast.iter_fields(statement):
-                if isinstance(value, list) and value and all(isinstance(item, ast.stmt) for item in value):
-                    visit_statements(value)
-
-    visit_statements(tree.body)
-    return endif_ranges
-
-
-def _module_string_constants(tree: ast.Module) -> dict[str, str]:
-    """Return top-level string assignments from a compiled Jinja module AST."""
-    constants: dict[str, str] = {}
-    for node in tree.body:
-        if not isinstance(node, ast.Assign) or not isinstance(node.value, ast.Constant) or not isinstance(node.value.value, str):
-            continue
-
-        for target in node.targets:
-            if isinstance(target, ast.Name):
-                constants[target.id] = node.value.value
-
-    return constants
-
-
-def _resolve_template_source_filename(template_name: str, compiled_root: Path) -> Path | None:
-    """
-    Resolve a Jinja template name from a compiled module to an existing source file.
-
-    The source template root is the parent of the configured
-    ``compiled_templates`` directory. The resolved path must stay within that
-    root to avoid trusting arbitrary paths from generated module constants.
-    """
-    template_root = compiled_root.parent.resolve()
-    source_filename = template_root.joinpath(*Path(template_name.replace("\\", "/")).parts).resolve()
-    if not source_filename.is_relative_to(template_root) or not source_filename.is_file():
-        return None
-
-    return source_filename
-
-
-def _parse_debug_info(debug_info: str) -> tuple[tuple[int, int], ...]:
-    """Parse Jinja's ``debug_info`` string into ``(generated_line, template_line)`` pairs."""
-    debug_map: list[tuple[int, int]] = []
-    for pair in debug_info.split("&"):
-        if not pair:
-            continue
-
-        try:
-            template_line, generated_line = (int(value) for value in pair.split("=", 1))
-        except ValueError:
-            continue
-
-        debug_map.append((generated_line, template_line))
-
-    return tuple(sorted(debug_map))
-
-
-def _find_reportable_jinja_lines(filename: Path) -> frozenset[int]:
-    """Return reportable source lines for a Jinja template with file-stamp cache invalidation."""
-    if (file_stamp := _file_stamp(filename)) is None:
-        return frozenset()
-
-    return _find_reportable_jinja_lines_cached(filename.resolve(), file_stamp)
-
-
-@lru_cache(maxsize=4096)
-def _find_reportable_jinja_lines_cached(filename: Path, file_stamp: FileStamp) -> frozenset[int]:  # noqa: ARG001
-    """
-    Parse a Jinja template and identify lines that represent source behavior.
-
-    Jinja's AST does not make static text output look like normal statements,
-    so reportable lines are a union of AST nodes, static data tokens, and
-    selected control-statement block starts. ``file_stamp`` participates only
-    in cache invalidation.
-    """
-    try:
-        source = filename.read_text(encoding="utf-8")
-    except (OSError, UnicodeDecodeError):
-        return frozenset()
-
-    try:
-        from jinja2 import Environment, TemplateSyntaxError, nodes  # noqa: PLC0415
-    except ImportError:
-        return frozenset()
-
-    environment = Environment(extensions=JINJA2_EXTENSIONS)  # noqa: S701
-    try:
-        parsed_template = environment.parse(source)
-    except TemplateSyntaxError:
-        return frozenset()
-
-    ignored_nodes = (nodes.Output, nodes.Template)
-    reportable_lines: set[int] = set()
-    nodes_to_visit = list(parsed_template.iter_child_nodes())
-    while nodes_to_visit:
-        node = nodes_to_visit.pop()
-        if isinstance(node, nodes.Stmt) and not isinstance(node, ignored_nodes):
-            lineno = getattr(node, "lineno", None)
-            if isinstance(lineno, int) and lineno > 0:
-                reportable_lines.add(lineno)
-
-        nodes_to_visit.extend(node.iter_child_nodes())
-
-    reportable_lines.update(line_number for line_number, _line_text in _static_template_lines(source))
-    reportable_lines = _expand_tag_ranges(source, reportable_lines | _variable_statement_lines(source) | _control_statement_lines(source))
-    return frozenset(reportable_lines)
-
-
-@lru_cache(maxsize=4096)
-def _find_structural_control_label_lines_cached(filename: Path, file_stamp: FileStamp) -> frozenset[int]:  # noqa: ARG001
-    """Return structural control labels used only as branch arc endpoints."""
-    try:
-        source = filename.read_text(encoding="utf-8")
-    except (OSError, UnicodeDecodeError):
-        return frozenset()
-
-    return frozenset(_control_statement_lines(source, names={"endif"}))
-
-
-def _generated_static_line_ranges(tree: ast.Module, source_filename: Path) -> dict[int, tuple[int, int]]:
-    """
-    Map generated ``yield 'static text'`` lines back to source template ranges.
-
-    Jinja ``debug_info`` does not reliably credit static output text to source
-    lines. This function uses generated literal-yield nodes as runtime evidence
-    and matches their non-empty rendered lines back to the static text tokens in
-    the source template. Ambiguous matches are ignored.
-    """
-    try:
-        source = source_filename.read_text(encoding="utf-8")
-    except (OSError, UnicodeDecodeError):
-        return {}
-
-    source_static_lines = _static_template_lines(source)
-    if not source_static_lines:
-        return {}
-
-    ranges: dict[int, tuple[int, int]] = {}
-    source_index = 0
-    yield_nodes = sorted(
-        (node for node in ast.walk(tree) if isinstance(node, ast.Yield)),
-        key=lambda node: node.lineno,
-    )
-    for node in yield_nodes:
-        if not isinstance(node, ast.Yield) or not isinstance(node.value, ast.Constant) or not isinstance(node.value.value, str):
-            continue
-
-        rendered_lines = _non_whitespace_lines(node.value.value)
-        if not rendered_lines:
-            continue
-
-        matched_line_numbers: list[int] = []
-        next_source_index = source_index
-        for rendered_line in rendered_lines:
-            matched_index = _find_next_static_line(source_static_lines, rendered_line, next_source_index)
-            if matched_index is None:
-                matched_line_numbers.clear()
-                break
-
-            matched_line_numbers.append(source_static_lines[matched_index][0])
-            next_source_index = matched_index + 1
-
-        if matched_line_numbers:
-            ranges[node.lineno] = (matched_line_numbers[0], matched_line_numbers[-1])
-            source_index = next_source_index
-
-    return ranges
-
-
-def _find_next_static_line(source_static_lines: tuple[tuple[int, str], ...], rendered_line: str, start_index: int) -> int | None:
-    """Find the next source static line matching one rendered static output line."""
-    for index, (_line_number, source_line) in enumerate(source_static_lines[start_index:], start=start_index):
-        if source_line == rendered_line:
-            return index
-
-    return None
-
-
-def _static_template_lines(source: str) -> tuple[tuple[int, str], ...]:
-    """Return non-empty static-data lines from a Jinja source template."""
-    try:
-        from jinja2 import Environment  # noqa: PLC0415
-    except ImportError:
-        return ()
-
-    environment = Environment(extensions=JINJA2_EXTENSIONS)  # noqa: S701
-    static_lines: list[tuple[int, str]] = []
-    try:
-        tokens = environment.lex(source)
-        for lineno, kind, value in tokens:
-            if kind == "data":
-                static_lines.extend(_numbered_non_whitespace_lines(lineno, value))
-    except Exception:
-        return ()
-
-    return tuple(static_lines)
-
-
-def _source_tag_ranges(source_filename: Path) -> dict[int, tuple[int, int]]:
-    """Return source line to full executable Jinja tag range mappings."""
-    try:
-        source = source_filename.read_text(encoding="utf-8")
-    except (OSError, UnicodeDecodeError):
-        return {}
-
-    return _tag_ranges_by_line(source)
-
-
-def _expand_tag_ranges(source: str, line_numbers: set[int]) -> set[int]:
-    """Expand any line inside a Jinja tag to the full tag source range."""
-    tag_ranges = _tag_ranges_by_line(source)
-    expanded_lines: set[int] = set()
-    for line_number in line_numbers:
-        start_line, end_line = tag_ranges.get(line_number, (line_number, line_number))
-        expanded_lines.update(range(start_line, end_line + 1))
-
-    return expanded_lines
-
-
-def _tag_ranges_by_line(source: str) -> dict[int, tuple[int, int]]:
-    """Return line mappings for block and variable tags, including multiline tags."""
-    try:
-        from jinja2 import Environment  # noqa: PLC0415
-    except ImportError:
-        return {}
-
-    environment = Environment(extensions=JINJA2_EXTENSIONS)  # noqa: S701
-    tag_ranges: dict[int, tuple[int, int]] = {}
-    tag_start_line: int | None = None
-    end_kinds: set[str] = set()
-    try:
-        for lineno, kind, _value in environment.lex(source):
-            if kind == "block_begin":
-                tag_start_line = lineno
-                end_kinds = {"block_end"}
-                continue
-
-            if kind == "variable_begin":
-                tag_start_line = lineno
-                end_kinds = {"variable_end"}
-                continue
-
-            if tag_start_line is not None and kind in end_kinds:
-                tag_range = (tag_start_line, lineno)
-                for line_number in range(tag_start_line, lineno + 1):
-                    tag_ranges[line_number] = tag_range
-                tag_start_line = None
-                end_kinds = set()
-    except Exception:
-        return {}
-
-    return tag_ranges
-
-
-def _else_branch_line_ranges(source_filename: Path) -> dict[int, tuple[int, int]]:
-    """Return first else-body reportable line to else-tag range mappings."""
-    try:
-        source = source_filename.read_text(encoding="utf-8")
-    except (OSError, UnicodeDecodeError):
-        return {}
-
-    try:
-        from jinja2 import Environment, TemplateSyntaxError, nodes  # noqa: PLC0415
-    except ImportError:
-        return {}
-
-    environment = Environment(extensions=JINJA2_EXTENSIONS)  # noqa: S701
-    try:
-        parsed_template = environment.parse(source)
-    except TemplateSyntaxError:
-        return {}
-
-    reportable_lines = _find_reportable_jinja_lines(source_filename)
-    tag_ranges = _tag_ranges_by_line(source)
-    else_lines = _control_statement_lines(source, names={"else"})
-    else_branch_line_ranges: dict[int, tuple[int, int]] = {}
-    for node in parsed_template.find_all((nodes.If, nodes.For)):
-        if not node.else_:
-            continue
-
-        first_else_node_line = min(getattr(child_node, "lineno", 0) for child_node in node.else_)
-        else_line = next((line for line in sorted(else_lines, reverse=True) if node.lineno < line <= first_else_node_line), None)
-        if else_line is None:
-            continue
-
-        body_line = _first_reportable_line_in_nodes(reportable_lines, node.else_, after_line=_tag_end_line(else_line, tag_ranges))
-        if body_line is not None:
-            else_branch_line_ranges[body_line] = (else_line, body_line)
-
-    return else_branch_line_ranges
-
-
-def _variable_statement_lines(source: str) -> set[int]:
-    """Return source lines for Jinja variable statements."""
-    try:
-        from jinja2 import Environment  # noqa: PLC0415
-    except ImportError:
-        return set()
-
-    environment = Environment(extensions=JINJA2_EXTENSIONS)  # noqa: S701
-    try:
-        return {lineno for lineno, kind, _value in environment.lex(source) if kind == "variable_begin"}
-    except Exception:
-        return set()
-
-
-def _control_statement_lines(source: str, names: set[str] | None = None) -> set[int]:
-    """
-    Return block-start lines for control statements that should be reportable.
-
-    Jinja may attach a node line to the expression inside a block rather than
-    to the ``{%`` line itself. Lexing preserves the block-start line, so this
-    helper adds the source line for ``if``, ``elif``, and ``for`` statements.
-    """
-    names = names or {"if", "elif", "else", "for"}
-    return {line_number for line_number, name in _block_statement_lines(source) if name in names}
-
-
-def _if_endif_lines(source: str) -> dict[int, int]:
-    """Return matching source ``endif`` lines for each ``if`` and ``elif`` line."""
-    endif_lines_by_conditional_line: dict[int, int] = {}
-    if_stack: list[list[int]] = []
-    for block_lineno, block_name in _block_statement_lines(source):
-        if block_name == "if":
-            if_stack.append([block_lineno])
-        elif block_name == "elif" and if_stack:
-            if_stack[-1].append(block_lineno)
-        elif block_name == "endif" and if_stack:
-            for conditional_line in if_stack.pop():
-                endif_lines_by_conditional_line[conditional_line] = block_lineno
-
-    return endif_lines_by_conditional_line
-
-
-def _block_statement_lines(source: str) -> tuple[tuple[int, str], ...]:
-    """Return the first statement name and source line for each Jinja block."""
-    try:
-        from jinja2 import Environment  # noqa: PLC0415
-    except ImportError:
-        return ()
-
-    environment = Environment(extensions=JINJA2_EXTENSIONS)  # noqa: S701
-    block_statements: list[tuple[int, str]] = []
-    in_block = False
-    block_lineno = 0
-    try:
-        for lineno, kind, value in environment.lex(source):
-            if kind == "block_begin":
-                in_block = True
-                block_lineno = lineno
-            elif in_block and kind == "name":
-                block_statements.append((block_lineno, value))
-                in_block = False
-            elif kind == "block_end":
-                in_block = False
-    except Exception:
-        return ()
-
-    return tuple(block_statements)
-
-
-def _numbered_non_whitespace_lines(start_lineno: int, value: str) -> list[tuple[int, str]]:
-    """Split a Jinja static-data token into stripped non-empty lines with source line numbers."""
-    numbered_lines: list[tuple[int, str]] = []
-    line_number = start_lineno
-    for line in value.splitlines(keepends=True):
-        line_without_newline = line.removesuffix("\n").removesuffix("\r")
-        stripped_line = line_without_newline.strip()
-        if stripped_line:
-            numbered_lines.append((line_number, stripped_line))
-
-        if line.endswith(("\n", "\r")):
-            line_number += 1
-
-    return numbered_lines
-
-
-def _non_whitespace_lines(value: str) -> list[str]:
-    """Return stripped non-empty lines from rendered static output."""
-    return [line.strip() for line in value.splitlines() if line.strip()]
-
-
-def _find_possible_jinja_arcs(source_filename: Path) -> frozenset[tuple[int, int]]:
-    """Return possible source-level branch arcs for a template with file-stamp cache invalidation."""
-    if (file_stamp := _file_stamp(source_filename)) is None:
-        return frozenset()
-
-    reportable_lines = _find_reportable_jinja_lines(source_filename)
-    structural_control_label_lines = _find_structural_control_label_lines_cached(source_filename.resolve(), file_stamp)
-    return _find_possible_jinja_arcs_cached(
-        source_filename.resolve(),
-        file_stamp,
-        reportable_lines,
-        _arc_endpoint_lines(reportable_lines, structural_control_label_lines),
-    )
-
-
-@lru_cache(maxsize=4096)
-def _find_possible_jinja_arcs_cached(
-    source_filename: Path,
-    file_stamp: FileStamp,  # noqa: ARG001
-    reportable_lines: frozenset[int],
-    arc_endpoint_lines: frozenset[int],
-) -> frozenset[tuple[int, int]]:
-    """
-    Parse supported Jinja control flow into possible source-level branch arcs.
-
-    The model intentionally covers only common source constructs where a
-    source-level branch is useful: ``if``/``elif``/``else`` and ``for`` with an
-    explicit ``else``. ``file_stamp`` is used only for cache invalidation.
-    """
-    try:
-        source = source_filename.read_text(encoding="utf-8")
-    except (OSError, UnicodeDecodeError):
-        return frozenset()
-
-    try:
-        from jinja2 import Environment, TemplateSyntaxError, nodes  # noqa: PLC0415
-    except ImportError:
-        return frozenset()
-
-    environment = Environment(extensions=JINJA2_EXTENSIONS)  # noqa: S701
-    try:
-        parsed_template = environment.parse(source)
-    except TemplateSyntaxError:
-        return frozenset()
-
-    tag_ranges = _tag_ranges_by_line(source)
-    endif_lines_by_conditional_line = _if_endif_lines(source)
-    elif_node_ids = {id(elif_node) for if_node in parsed_template.find_all(nodes.If) for elif_node in if_node.elif_}
-    arcs: set[tuple[int, int]] = set()
-    for node in parsed_template.find_all((nodes.If, nodes.For)):
-        if isinstance(node, nodes.If):
-            if id(node) in elif_node_ids:
-                continue
-            _add_if_arcs(node, reportable_lines, arc_endpoint_lines, tag_ranges, endif_lines_by_conditional_line, arcs)
-        elif isinstance(node, nodes.For):
-            _add_for_arcs(node, reportable_lines, arc_endpoint_lines, tag_ranges, arcs)
-
-    return frozenset(arcs)
-
-
-def _add_if_arcs(
-    node: Any,
-    reportable_lines: Collection[int],
-    arc_endpoint_lines: Collection[int],
-    tag_ranges: Mapping[int, tuple[int, int]],
-    endif_lines_by_conditional_line: Mapping[int, int],
-    arcs: set[tuple[int, int]],
-) -> None:
-    """Add possible arcs from a Jinja ``if`` node and any ``elif`` nodes."""
-    after_line = _next_reportable_line(reportable_lines, _node_end_lineno(node))
-    conditional_nodes = [node, *node.elif_]
-    for index, conditional_node in enumerate(conditional_nodes):
-        conditional_end_line = _tag_end_line(conditional_node.lineno, tag_ranges)
-        body_line = _first_reportable_line_in_nodes(
-            reportable_lines,
-            conditional_node.body,
-            after_line=conditional_end_line,
-        )
-        false_line = _false_if_target(node, conditional_nodes, index, reportable_lines, endif_lines_by_conditional_line, after_line)
-
-        _add_arc(arcs, conditional_node.lineno, body_line, arc_endpoint_lines)
-        _add_arc(arcs, conditional_node.lineno, false_line, arc_endpoint_lines)
-
-
-def _false_if_target(
-    node: Any,
-    conditional_nodes: list[Any],
-    index: int,
-    reportable_lines: Collection[int],
-    endif_lines_by_conditional_line: Mapping[int, int],
-    after_line: int,
-) -> int | None:
-    """Return the source line reached when one ``if`` or ``elif`` condition is false."""
-    if index + 1 < len(conditional_nodes):
-        return conditional_nodes[index + 1].lineno
-    if node.else_:
-        return _first_reportable_line_in_nodes(reportable_lines, node.else_)
-
-    return endif_lines_by_conditional_line.get(conditional_nodes[index].lineno, after_line)
-
-
-def _add_for_arcs(
-    node: Any,
-    reportable_lines: Collection[int],
-    arc_endpoint_lines: Collection[int],
-    tag_ranges: Mapping[int, tuple[int, int]],
-    arcs: set[tuple[int, int]],
-) -> None:
-    """Add possible arcs from a Jinja ``for`` node to its body and explicit ``else`` block."""
-    body_line = _first_reportable_line_in_nodes(reportable_lines, node.body, after_line=_tag_end_line(node.lineno, tag_ranges))
-
-    _add_arc(arcs, node.lineno, body_line, arc_endpoint_lines)
-    if node.else_:
-        _add_arc(arcs, node.lineno, _first_reportable_line_in_nodes(reportable_lines, node.else_), arc_endpoint_lines)
-
-
-def _first_reportable_line_in_nodes(reportable_lines: Collection[int], nodes, after_line: int = 0) -> int | None:  # noqa: ANN001
-    """Return the first reportable line associated with a list of Jinja AST nodes."""
-    if not nodes:
-        return None
-
-    start_line = min(getattr(node, "lineno", 0) for node in nodes)
-    end_line = max(_node_end_lineno(node) for node in nodes)
-    if line := next((line for line in sorted(reportable_lines) if line > after_line and start_line <= line <= end_line), None):
-        return line
-
-    return _next_reportable_line(reportable_lines, max(after_line, start_line - 1))
-
-
-def _next_reportable_line(reportable_lines: Collection[int], line_number: int) -> int:
-    """Return the next reportable line after ``line_number`` or the source-exit sentinel."""
-    return next((line for line in sorted(reportable_lines) if line > line_number), SOURCE_EXIT)
-
-
-def _tag_end_line(line_number: int, tag_ranges: Mapping[int, tuple[int, int]]) -> int:
-    """Return the end line of the Jinja tag containing ``line_number``."""
-    return tag_ranges.get(line_number, (line_number, line_number))[1]
-
-
-def _node_end_lineno(node) -> int:  # noqa: ANN001
-    """Return the greatest source line used by a Jinja AST node and its children."""
-    node_lines = [getattr(node, "lineno", 0)]
-    nodes_to_visit = list(node.iter_child_nodes())
-    while nodes_to_visit:
-        child_node = nodes_to_visit.pop()
-        if isinstance((lineno := getattr(child_node, "lineno", None)), int):
-            node_lines.append(lineno)
-        nodes_to_visit.extend(child_node.iter_child_nodes())
-
-    return max(node_lines)
-
-
-def _add_arc(arcs: set[tuple[int, int]], from_line: int, to_line: int | None, arc_endpoint_lines: Collection[int]) -> None:
-    """Add a branch arc when both endpoints belong to the source endpoint model."""
-    if to_line is None or from_line == to_line:
-        return
-
-    if from_line not in arc_endpoint_lines or (to_line > 0 and to_line not in arc_endpoint_lines):
-        return
-
-    arcs.add((from_line, to_line))
-
-
-def _find_no_branch_jinja_lines(source_filename: Path) -> frozenset[int]:
-    """Return source lines that should be excused from branch coverage with cache invalidation."""
-    if (file_stamp := _file_stamp(source_filename)) is None:
-        return frozenset()
-
-    return _find_no_branch_jinja_lines_cached(source_filename.resolve(), file_stamp)
-
-
-@lru_cache(maxsize=4096)
-def _find_no_branch_jinja_lines_cached(source_filename: Path, file_stamp: FileStamp) -> frozenset[int]:  # noqa: ARG001
-    """
-    Find top-level optional guards that should not count as partial branches.
-
-    Many AVD templates wrap optional feature sections in a top-level ``if``
-    without an ``elif`` or ``else``. Reporting the absent path as a missing
-    branch creates broad noise, so these simple guards are marked no-branch.
-    ``file_stamp`` is used only for cache invalidation.
-    """
-    try:
-        source = source_filename.read_text(encoding="utf-8")
-    except (OSError, UnicodeDecodeError):
-        return frozenset()
-
-    try:
-        from jinja2 import Environment, TemplateSyntaxError, nodes  # noqa: PLC0415
-    except ImportError:
-        return frozenset()
-
-    environment = Environment(extensions=JINJA2_EXTENSIONS)  # noqa: S701
-    try:
-        parsed_template = environment.parse(source)
-    except TemplateSyntaxError:
-        return frozenset()
-
-    return frozenset(node.lineno for node in parsed_template.body if isinstance(node, nodes.If) and not node.elif_ and not node.else_)
-
-
-def _file_stamp(filename: Path) -> FileStamp | None:
-    """Return the file metadata used to invalidate path-keyed caches."""
-    try:
-        stat_result = filename.stat()
-    except OSError:
-        return None
-
-    return stat_result.st_mtime_ns, stat_result.st_size
